move DB connection function into its own file
To avoid loading create_app libraries when creating the
database connection URL, move the URL generation code
into its own file. This is useful when performing Alembic
DB upgrades, as the update needs to occur without loading
any flask app code.
